+++ internal/bloblang/query/functions.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"io/ioutil"
 	"math/rand"
 	"os"
 	"time"
@@ -91,7 +92,7 @@ func NewLiteralFunction(v interface{}) *Literal {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "batch_index",
 		"Returns the index of the mapped message within a batch. This is useful for applying maps only on certain messages of a batch.",
@@ -108,7 +109,7 @@ var _ = RegisterFunctionSpec(
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "batch_size",
 		"Returns the size of the message batch.",
@@ -125,7 +126,7 @@ var _ = RegisterFunctionSpec(
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "content",
 		"Returns the full raw contents of the mapping target message as a byte array. When mapping to a JSON field the value should be encoded using the method [`encode`][methods.encode], or cast to a string directly using the method [`string`][methods.string], otherwise it will be base64 encoded by default.",
@@ -146,7 +147,7 @@ func contentFunction(...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "count",
 		"The `count` function is a counter starting at 1 which increments after each time it is called. Count takes an argument which is an identifier for the counter, allowing you to specify multiple unique counters in your configuration.",
@@ -187,7 +188,7 @@ func countFunction(args ...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "deleted",
 		"A function that returns a result indicating that the mapping target should be deleted.",
@@ -216,7 +217,7 @@ root.bar = deleted()`,
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "env",
 		"Returns the value of an environment variable.",
@@ -230,15 +231,13 @@ var _ = RegisterFunctionSpec(
 )
 
 func envFunction(args ...interface{}) (Function, error) {
-	key := args[0].(string)
-	return ClosureFunction(func(_ FunctionContext) (interface{}, error) {
-		return os.Getenv(key), nil
-	}, nil), nil
+	key := os.Getenv(args[0].(string))
+	return NewLiteralFunction(key), nil
 }
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "error",
 		"If an error has occurred during the processing of a message this function returns the reported cause of the error. For more information about error handling patterns read [here][error_handling].",
@@ -255,7 +254,7 @@ func errorFunction(...interface{}) (Function, error) {
 	}, nil), nil
 }
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "errored",
 		"Returns a boolean value indicating whether an error has occurred during the processing of a message. For more information about error handling patterns read [here][error_handling].",
@@ -274,7 +273,32 @@ func erroredFunction(...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
+	NewFunctionSpec(
+		FunctionCategoryEnvironment, "file",
+		"Reads a file and returns its contents. Relative paths are resolved from the directory of the process executing the mapping.",
+		NewExampleSpec("",
+			`root.doc = file(env("BENTHOS_TEST_BLOBLANG_FILE")).parse_json()`,
+			`{}`,
+			`{"doc":{"foo":"bar"}}`,
+		),
+	).IsBeta(true),
+	true, fileFunction,
+	ExpectNArgs(1),
+	ExpectStringArg(0),
+)
+
+func fileFunction(args ...interface{}) (Function, error) {
+	pathBytes, err := ioutil.ReadFile(args[0].(string))
+	if err != nil {
+		return nil, err
+	}
+	return NewLiteralFunction(pathBytes), nil
+}
+
+//------------------------------------------------------------------------------
+
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "range",
 		"The `range` function creates an array of integers following a range between a start, stop and optional step integer argument. If the step argument is omitted then it defaults to 1. A negative step can be provided as long as stop < start.",
@@ -312,7 +336,7 @@ func rangeFunction(args ...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "hostname",
 		"Returns a string matching the hostname of the machine running Benthos.",
@@ -338,7 +362,7 @@ func hostnameFunction(...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "json",
 		"Returns the value of a field within a JSON message located by a [dot path][field_paths] argument. This function always targets the entire source JSON document regardless of the mapping context.",
@@ -386,7 +410,7 @@ func jsonFunction(args ...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "meta",
 		"Returns the value of a metadata key from a message identified by a key. Values are extracted from the referenced input message and therefore do NOT reflect changes made from within the map.",
@@ -439,7 +463,7 @@ func metadataFunction(args ...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewDeprecatedFunctionSpec("nothing"),
 	false, func(...interface{}) (Function, error) {
 		return NewLiteralFunction(Nothing(nil)), nil
@@ -448,7 +472,7 @@ var _ = RegisterFunctionSpec(
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "random_int",
 		"Generates a non-negative pseudo-random 64-bit integer. An optional integer argument can be provided in order to seed the random number generator.",
@@ -478,7 +502,7 @@ func randomIntFunction(args ...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "timestamp",
 		"Returns the current time in a custom format specified by the argument. The format is defined by showing how the reference time, defined to be `Mon Jan 2 15:04:05 -0700 MST 2006` would be displayed if it were the value.\n\nA fractional second is represented by adding a period and zeros to the end of the seconds section of layout string, as in `15:04:05.000` to format a time stamp with millisecond precision.",
@@ -499,7 +523,7 @@ var _ = RegisterFunctionSpec(
 	ExpectStringArg(0),
 )
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "timestamp_utc",
 		"The equivalent of `timestamp` except the time is printed as UTC instead of the local timezone.",
@@ -520,7 +544,7 @@ var _ = RegisterFunctionSpec(
 	ExpectStringArg(0),
 )
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "timestamp_unix",
 		"Returns the current unix timestamp in seconds.",
@@ -535,7 +559,7 @@ var _ = RegisterFunctionSpec(
 	},
 )
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryEnvironment, "timestamp_unix_nano",
 		"Returns the current unix timestamp in nanoseconds.",
@@ -552,7 +576,7 @@ var _ = RegisterFunctionSpec(
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "throw",
 		"Throws an error similar to a regular mapping error. This is useful for abandoning a mapping entirely given certain conditions.",
@@ -581,7 +605,7 @@ root.doc.contents = (this.body.content | this.thing.body)`,
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryGeneral, "uuid_v4",
 		"Generates a new RFC-4122 UUID each time it is invoked and prints a string representation.",
@@ -602,7 +626,7 @@ func uuidFunction(...interface{}) (Function, error) {
 
 //------------------------------------------------------------------------------
 
-var _ = RegisterFunctionSpec(
+var _ = RegisterFunction(
 	NewDeprecatedFunctionSpec("var"), true, varFunction,
 	ExpectNArgs(1),
 	ExpectStringArg(0),